AccumulateOperands, PrivilegedServiceIndices.Copy for Kai_g (#637)
* AccumulateOperands encoding: ordering, gas uint
* PrivilegedServiceIndices.Copy
* updated algo + auth_copy for matching parse_accumulate_operands

TODO:
* update bootstrap, check fib; test in fuzzer
* increase iterations in algo

+++ node/spec_test.go
@@ -249,7 +249,7 @@ func TestBootstrapCodeFromSpec(t *testing.T) {
 			A: p_a,
 			O: r.Ok,
 			Y: result.PayloadHash,
-			G: result.Gas, // REVIEW
+			G: uint(result.Gas),
 			D: result.Result,
 		}
 		fmt.Printf("o: %x res: %s\n", o, result.String())
